\name{andechs}
\docType{data}
\alias{andechs}
\title{Fen Meadows}
\description{
This data set gives the average cover of vascular plant
species in subplots nested within 17 whole-plots from
mown fen meadows. This is a subset of the data used in
Schmidtlein & Sassin (2004).}
\usage{data(andechs)}
\format{A matrix containing 17 observations and 110
species.}
\source{
Schmidtlein, S., Sassin, J. (2004): Mapping of continuous
floristic gradients in grasslands using hyperspectral
imagery. \emph{Remote Sensing of Environment} \bold{92},
126--138.}
\keyword{datasets}